Custom TiniElement finalizeStyles()
