+++ apps/tinijs.dev/app/public/benchmark-reports/hello-world.json
@@ -1,1 +1,1 @@
-{"1":{"score":1,"fcp":"1.3 s","lcp":"1.5 s","tbt":"0 ms","cls":"0","si":"1.3 s"}}
+{"1":{"score":1,"fcp":"1.2 s","lcp":"1.4 s","tbt":"0 ms","cls":"0","si":"1.2 s"}}
