+++ apps/tinijs.dev/app/public/benchmark-reports/ui-text.json
@@ -1,1 +1,1 @@
-{"10":{"score":1,"fcp":"1.2 s","lcp":"1.4 s","tbt":"0 ms","cls":"0","si":"1.2 s"},"1000":{"score":1,"fcp":"1.2 s","lcp":"1.4 s","tbt":"60 ms","cls":"0","si":"1.2 s"},"10000":{"score":0.71,"fcp":"1.2 s","lcp":"1.5 s","tbt":"2,560 ms","cls":"0","si":"1.8 s"}}
+{"10":{"score":1,"fcp":"1.2 s","lcp":"1.4 s","tbt":"0 ms","cls":"0","si":"1.2 s"},"1000":{"score":1,"fcp":"1.2 s","lcp":"1.4 s","tbt":"60 ms","cls":"0","si":"1.2 s"},"10000":{"score":0.71,"fcp":"1.2 s","lcp":"1.5 s","tbt":"2,550 ms","cls":"0","si":"1.8 s"}}
