+++ apps/tinijs.dev/content/framework-posts/103 - benchmark-limit/index.md
@@ -61,7 +61,7 @@ Notes:
       <th>Framework</th>
       <th>App size</th>
       <th>Score</th>
-      <th>Speed Index</th>
+      <th>FCP</th>
       <th>LCP</th>
     </tr>
   </thead>
@@ -73,7 +73,7 @@ Notes:
         <strong>Vanilla</strong> <br>
         <em>(Vite <a target="_blank" href="https://vitejs.dev/guide/">vanilla-ts</a>)</em>
       </td>
-      <td>1.8 kB</td>
+      <td>2 kB</td>
       <td><span class="ok">100</span></td>
       <td><span class="ok">0.8 s</span></td>
       <td><span class="ok">0.8 s</span></td>
@@ -84,7 +84,7 @@ Notes:
         <strong>Lit</strong> 3.1.2 <br>
         <em>(Vite <a target="_blank" href="https://vitejs.dev/guide/">lit-ts</a>)</em>
       </td>
-      <td>18.3 kB</td>
+      <td>18 kB</td>
       <td><span class="ok">100</span></td>
       <td><span class="ok">0.9 s</span></td>
       <td><span class="ok">0.9 s</span></td>
@@ -95,7 +95,7 @@ Notes:
         <strong>Tini</strong> 0.20.0 <br>
         <em>(App only - <a target="_blank" href="https://github.com/tinijs/bare-starter">Bare</a> starter)</em>
       </td>
-      <td>37.4 kB</td>
+      <td>37 kB</td>
       <td><span class="ok">100</span></td>
       <td><span class="ok">0.9 s</span></td>
       <td><span class="ok">0.9 s</span></td>
@@ -106,9 +106,9 @@ Notes:
         <strong>Tini</strong> 0.20.0 <br>
         <em>(App, Router, Meta and UI - <a target="_blank" href="https://github.com/tinijs/blank-starter">Blank</a> starter)</em>
       </td>
-      <td>74.2 kB</td>
+      <td>74 kB</td>
       <td><span class="ok">100</span></td>
-      <td><span class="ok">1.3 s</span></td>
+      <td><span class="ok">1.2 s</span></td>
       <td><span class="ok">1.7 s</span></td>
     </tr>
 
@@ -117,7 +117,7 @@ Notes:
         <strong>Vue</strong> 3.4.21 <br>
         <em>(Vite <a target="_blank" href="https://vitejs.dev/guide/">vue-ts</a>)</em>
       </td>
-      <td>54.4 kB</td>
+      <td>54 kB</td>
       <td><span class="ok">100</span></td>
       <td><span class="ok">1.1 s</span></td>
       <td><span class="ok">1.1 s</span></td>
@@ -128,7 +128,7 @@ Notes:
         <strong>React</strong> 18.2.0 <br>
         <em>(Vite <a target="_blank" href="https://vitejs.dev/guide/">react-ts</a>)</em>
       </td>
-      <td>143.7 kB</td>
+      <td>144 kB</td>
       <td><span class="ok">100</span></td>
       <td><span class="ok">1.5 s</span></td>
       <td><span class="ok">1.5 s</span></td>
@@ -139,7 +139,7 @@ Notes:
         <strong>Angular</strong> 18.0.0 <br>
         <em>(Angular <a target="_blank" href="https://angular.dev/tools/cli/setup-local">ng new</a>)</em>
       </td>
-      <td>212.3 kB</td>
+      <td>212 kB</td>
       <td><span class="ok">98</span></td>
       <td><span class="md">2.0 s</span></td>
       <td><span class="ok">2.0 s</span></td>
@@ -150,7 +150,7 @@ Notes:
         <strong>Svelte</strong> 4.2.12 <br>
         <em>(Vite <a target="_blank" href="https://vitejs.dev/guide/">svelte-ts</a>)</em>
       </td>
-      <td>4.7 kB</td>
+      <td>5 kB</td>
       <td><span class="ok">100</span></td>
       <td><span class="ok">0.8 s</span></td>
       <td><span class="ok">0.8 s</span></td>
@@ -161,7 +161,7 @@ Notes:
         <strong>Solid</strong> 1.8.15 <br>
         <em>(Vite <a target="_blank" href="https://vitejs.dev/guide/">solid-ts</a>)</em>
       </td>
-      <td>7.6 kB</td>
+      <td>8 kB</td>
       <td><span class="ok">100</span></td>
       <td><span class="ok">0.8 s</span></td>
       <td><span class="ok">0.8 s</span></td>
